Refactor AuditResults.
closes #1965
