+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config --auto-gen-only-exclude`
-# on 2022-10-05 15:32:47 UTC using RuboCop version 1.36.0.
+# on 2022-10-20 12:10:43 UTC using RuboCop version 1.36.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -53,13 +53,12 @@ Metrics/BlockNesting:
   Exclude:
     - 'app/services/complete_moab_handler.rb'
 
-# Offense count: 5
+# Offense count: 4
 # Configuration parameters: CountComments, Max, CountAsOne.
 Metrics/ClassLength:
   Exclude:
     - 'app/controllers/objects_controller.rb'
     - 'app/models/druid_version_zip.rb'
-    - 'app/services/audit_results.rb'
     - 'app/services/checksum_validator.rb'
     - 'app/services/complete_moab_handler.rb'
 
@@ -80,15 +79,12 @@ Metrics/PerceivedComplexity:
   Exclude:
     - 'app/services/complete_moab_handler.rb'
 
-# Offense count: 9
+# Offense count: 3
 # Configuration parameters: EnforcedStyle, CheckMethodNames, CheckSymbols, AllowedIdentifiers, AllowedPatterns.
 # SupportedStyles: snake_case, normalcase, non_integer
 # AllowedIdentifiers: capture3, iso8601, rfc1123_date, rfc822, rfc2822, rfc3339
 Naming/VariableNumber:
   Exclude:
-    - 'spec/jobs/catalog_to_moab_job_spec.rb'
-    - 'spec/jobs/moab_to_catalog_job_spec.rb'
-    - 'spec/services/complete_moab_handler_check_exist_spec.rb'
     - 'spec/services/moab_storage_root_reporter_spec.rb'
 
 # Offense count: 1
@@ -107,12 +103,12 @@ RSpec/LetSetup:
   Exclude:
     - 'spec/services/complete_moab_handler_confirm_version_spec.rb'
 
-# Offense count: 416
+# Offense count: 407
 # Configuration parameters: AllowSubject.
 RSpec/MultipleMemoizedHelpers:
   Max: 21
 
-# Offense count: 20
+# Offense count: 18
 RSpec/StubbedMock:
   Exclude:
     - 'spec/jobs/catalog_to_moab_job_spec.rb'
