+++ app/services/audit/catalog_to_moab.rb
@@ -16,7 +16,7 @@ module Audit
       @complete_moab = complete_moab
       @druid = complete_moab.preserved_object.druid
       @logger = Logger.new(Rails.root.join('log', 'c2m.log'))
-      @results = AuditResults.new(druid, nil, complete_moab.moab_storage_root, logger: @logger)
+      @results = AuditResults.new(druid: druid, moab_storage_root: complete_moab.moab_storage_root)
     end
 
     # shameless green implementation
@@ -26,7 +26,7 @@ module Audit
         results.add_result(AuditResults::CM_PO_VERSION_MISMATCH,
                            cm_version: complete_moab.version,
                            po_version: complete_moab.preserved_object.current_version)
-        return results.report_results
+        return report_results!
       end
 
       unless online_moab_found?
@@ -39,15 +39,19 @@ module Audit
         results.add_result(AuditResults::MOAB_NOT_FOUND,
                            db_created_at: complete_moab.created_at.iso8601,
                            db_updated_at: complete_moab.updated_at.iso8601)
-        return results.report_results
+        return report_results!
       end
 
-      return results.report_results unless can_validate_current_comp_moab_status?
+      return report_results! unless can_validate_current_comp_moab_status?
 
       compare_version_and_take_action
     end
 
     private
+
+    def report_results!
+      AuditResultsReporter.report_results(audit_results: results, logger: @logger)
+    end
 
     def moab_validator
       @moab_validator ||= MoabValidator.new(druid: druid, storage_location: storage_location, results: results, complete_moab: complete_moab)
@@ -72,7 +76,7 @@ module Audit
         if catalog_version == moab_version
           set_status_as_seen_on_disk(true) unless complete_moab.ok?
           results.add_result(AuditResults::VERSION_MATCHES, 'CompleteMoab')
-          results.report_results
+          report_results!
         elsif catalog_version < moab_version
           set_status_as_seen_on_disk(true)
           comp_moab_handler = CompleteMoabHandler.new(druid, moab_version, moab.size, complete_moab.moab_storage_root)
@@ -82,7 +86,7 @@ module Audit
           results.add_result(
             AuditResults::UNEXPECTED_VERSION, db_obj_name: 'CompleteMoab', db_obj_version: complete_moab.version
           )
-          results.report_results
+          report_results!
         end
 
         complete_moab.update_audit_timestamps(ran_moab_validation?, true)
