+++ spec/services/checksum_validator_spec.rb
@@ -12,7 +12,7 @@ RSpec.describe ChecksumValidator do
   end
   let(:checksum_validator) { described_class.new(complete_moab) }
   let(:moab_validator) { checksum_validator.send(:moab_validator) }
-  let(:results) { instance_double(AuditResults, report_results: nil, check_name: nil, :actual_version= => nil) }
+  let(:results) { instance_double(AuditResults, check_name: nil, :actual_version= => nil) }
   let(:logger_double) { instance_double(ActiveSupport::Logger, info: nil, error: nil, add: nil) }
   let(:audit_workflow_reporter) { instance_double(Reporters::AuditWorkflowReporter, report_errors: nil, report_completed: nil) }
   let(:honeybadger_reporter) { instance_double(Reporters::HoneybadgerReporter, report_errors: nil, report_completed: nil) }
@@ -38,7 +38,7 @@ RSpec.describe ChecksumValidator do
       it 'sets status to online_moab_not_found and adds corresponding audit result' do
         expect { checksum_validator.validate_checksums }.to change(complete_moab, :status).to 'online_moab_not_found'
         expect(complete_moab.reload.status).to eq 'online_moab_not_found'
-        expect(checksum_validator.results.result_array.first).to have_key(:moab_not_found)
+        expect(checksum_validator.results.results.first).to have_key(:moab_not_found)
       end
 
       it 'sends results in HONEYBADGER_REPORT_CODES errors' do
@@ -62,7 +62,7 @@ RSpec.describe ChecksumValidator do
       end
 
       it 'calls AuditResults.report_results' do
-        expect(checksum_validator.results).to receive(:report_results)
+        expect(AuditResultsReporter).to receive(:report_results).with(audit_results: AuditResults, logger: Audit::Checksum.logger)
         checksum_validator.validate_checksums
       end
     end
@@ -78,7 +78,7 @@ RSpec.describe ChecksumValidator do
         expect(moab_validator.moab.object_pathname.exist?).to be true
         expect { checksum_validator.validate_checksums }.to change(complete_moab, :status).to 'online_moab_not_found'
         expect(complete_moab.reload.status).to eq 'online_moab_not_found'
-        expect(checksum_validator.results.result_array.first).to have_key(:moab_not_found)
+        expect(checksum_validator.results.results.first).to have_key(:moab_not_found)
       end
 
       it 'sends results in HONEYBADGER_REPORT_CODES errors' do
@@ -101,8 +101,8 @@ RSpec.describe ChecksumValidator do
                           { cm_status_changed: 'CompleteMoab status changed from validity_unknown to online_moab_not_found' }])
       end
 
-      it 'calls AuditResults.report_results' do
-        expect(checksum_validator.results).to receive(:report_results)
+      it 'calls AuditResultReporter.report_results' do
+        expect(AuditResultsReporter).to receive(:report_results).with(audit_results: AuditResults, logger: Audit::Checksum.logger)
         checksum_validator.validate_checksums
       end
     end
@@ -113,7 +113,7 @@ RSpec.describe ChecksumValidator do
 
       it 'returns a positive result for a complete_moab' do
         checksum_validator.validate_checksums
-        expect(checksum_validator.results.result_array.first).to have_key(:moab_checksum_valid)
+        expect(checksum_validator.results.results.first).to have_key(:moab_checksum_valid)
       end
 
       [
@@ -209,7 +209,7 @@ RSpec.describe ChecksumValidator do
     context 'fails checksum validation' do
       it 'returns error codes for a complete_moab' do
         checksum_validator.validate_checksums
-        expect(checksum_validator.results.result_array.first).to have_key(:file_not_in_manifest)
+        expect(checksum_validator.results.results.first).to have_key(:file_not_in_manifest)
       end
 
       [
@@ -290,7 +290,7 @@ RSpec.describe ChecksumValidator do
 
     context 'reports results' do
       it 'calls AuditResults.report_results' do
-        expect(checksum_validator.results).to receive(:report_results)
+        expect(AuditResultsReporter).to receive(:report_results).with(audit_results: AuditResults, logger: Audit::Checksum.logger)
         checksum_validator.validate_checksums
       end
     end
